+++ src/lib/acme-spring/src/main/resources/META-INF/services/org.shredzone.acme4j.provider.AcmeProvider
@@ -1,0 +1,1 @@
+org.niis.xroad.common.acme.spring.AcmeProfileIdProvider
