+++ src/security-server/admin-service/application/src/main/java/org/niis/xroad/securityserver/restapi/RestApiApplication.java
@@ -27,6 +27,7 @@ package org.niis.xroad.securityserver.restapi;
 
 import ee.ria.xroad.common.Version;
 
+import org.niis.xroad.globalconf.spring.GlobalConfBeanLookup;
 import org.niis.xroad.globalconf.spring.SpringGlobalConfConfig;
 import org.niis.xroad.globalconf.spring.SpringOcspVerifierConfig;
 import org.niis.xroad.serverconf.spring.ServerConfBeanConfig;
@@ -42,6 +43,7 @@ import org.springframework.context.annotation.Import;
 @Import({
         SpringGlobalConfConfig.class,
         SpringOcspVerifierConfig.class,
+        GlobalConfBeanLookup.class,
         ServerConfBeanConfig.class})
 @ServletComponentScan
 @SpringBootApplication(scanBasePackages = {"org.niis.xroad.securityserver.restapi", "org.niis.xroad.restapi", "org.niis.xroad.common.acme",
