--- src/security-server/admin-service/application/src/main/resources/META-INF/services/org.shredzone.acme4j.provider.AcmeProvider
@@ -1,2 +1,0 @@
-org.niis.xroad.securityserver.restapi.acme.AcmeXroadProvider
-org.niis.xroad.securityserver.restapi.acme.AcmeProfileIdProvider
